requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.0.0
pandas>=1.5.0
plotly>=5.15.0
huggingface-hub>=0.17.0
accelerate>=0.24.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0